requirements.txt
# =============================================================================
# Runtime Dependencies
# =============================================================================

# Agent framework
google-adk>=1.0.0

# Configuration
pyyaml>=6.0
pydantic>=2.0

# Environment
python-dotenv>=1.0

# Templating and content processing
jinja2>=3.1
markdown>=3.5
nh3>=0.2

# Gmail API
google-api-python-client>=2.100
google-auth>=2.20
google-auth-oauthlib>=1.2

# HTTP client (Perplexity API + E2E tests)
httpx>=0.27

# Observability
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-grpc>=1.20.0

# Perplexity API (OpenAI-compatible client)
openai>=1.0

# Cloud Run HTTP server
flask>=3.0
gunicorn>=22.0

# =============================================================================
# Test Dependencies
# =============================================================================
pytest>=8.0
pytest-asyncio>=0.23
pytest-cov>=5.0
